refactored gelf handler + testing for correct handling of duplicate entries
